Added an Arithmetic (Range) coder.
This is using the 32-bit version with carry correction, designed by
Michael Schindler (who based it on Graham Martin's paper) but as
reformulated by Eugene Shelwien.

The avoidance of 64-bit is key for Javascript performance and has
negligible impact on the C implementation, so it seems we should be
using this instead.

Decompression performance on q40b test is ~6.7s decode Order-0 vs 3.1s
in C.