Updated the FQZcomp parameter table.
The pseudocode still needs revising (as does the actual code).  This
defines multiple parameter blocks and an optional arbitrary selector
to distinguish between (read from model_sel if required).  The
selector bits do not have to be part of the context, although this is
probably recommended for efficient context construction.  To provide
options here, the selector bits are not explicitly placed into the
context and instead each parameter block defines a starting point for
the context value at the start of each new record.