Updated fqzcomp to handle multiple parameters (as per revised spec).
For now the encoder just uses a single one, but there is (not checked
in) a fqzcomp.q4+40.js variant hard coded to 151bp NovaSeq and 100bp
HiSeq mingled runs with length based detection for the selector.  This
is simply a proof of concept to validate the stream can be decoded.