Add READ1/READ2 support into fqzcomp via selector.
The main_fqzcomp demo program reads either

<quals>
<quals>
<quals>

or

<quals> [01]
<quals> [01]
<quals> [01]

When the latter is used, we allocate a bit of context for this (eg
READ2 status) and adjust the parameters accordingly. Mostly it's
minor, but on some data sets it can be significant.  This is mainly a
proof of the decoder working than a great set of encoding parameters.