Document the 4-way vs 32-way 16-bit rANS entropy encoder.
This bit controls the number of rANS states to rotate through during
the decode loop.  If unset it is 4, as in the old rANS4x8 codec, but
use of 32 states offers more opportunity for use of SIMD instructions.

The C htscodecs implementation demonstrates 2-3 fold speed increase is
possible, although sadly it is not something that compilers are able
to auto-vectorise.